Thêm chức năng: sửa xoá comment, sửa lại giao diện đọc truyện

+++ App/app/src/main/java/com/fpoly/assigment_mob403/DTO/Comment.java
@@ -57,7 +57,7 @@ public class Comment implements Serializable {
         return time;
     }
 
-    public void setTime(int time) {
+    public void setTime(long time) {
         this.time = time;
     }
